Rework FTB Omnia 1.0.0 advisory
It was suggested that the wording of this advisory was such that it makes the
site feel "political". The wording here was strong, probably too strong -
this is a more toned-down message, that encourages players to makeup their
own mind.

The purpose of this system is not to make choices for players, but rather to
make sure that players are aware of potentially game-breaking or ruining bugs
for (largely) old pack versions.

Having discussed the system with others, the instances that advisories appear
on the pack's page will now be significantly reduced. I am still deliberating
what criteria needs to be met (though I will likely omit beta versions, and
versions that aren't getting significant plays any more). Exactly how they are
written needs to be documented too.

Though it should be said that as of this point, our Advisory system is under
review. Concerns that the system could be abused or misused, are not ones I
want hanging over the project.

The original idea behind our additional metadata was for community
collaboration. At this time I am the only contributor, so there is a lack of
community oversight and input.

As a final point, I would like to use an official FTB-provided system here if
possible. I will be making an issue shortly for this.

Signed-off-by: Jamie Mansfield <jmansfield@cadixdev.org>